+++ sw/src/ibvStructs.cpp
@@ -30,7 +30,7 @@ void ibvQ::uintToGid(int idx, uint32_t ip_addr) {
 
 void ibvQ::print(const char *name) {
     printf("%s: QPN 0x%06x, PSN 0x%06x, VADDR %016lx, SIZE %08x, IP 0x%08x,\n",
-         name, qpn, psn, vaddr, size, ip_addr);
+         name, qpn, psn, (uint64_t)vaddr, size, ip_addr);
 }
 
 ibvQpPool::ibvQpPool(int32_t n_el) {
